mis en place les squelettes des appels d'api de lecture, creation et mise à jour d'utilisateur dans le controleur

+++ src/main/java/fr/lyon/insa/ot/sims/shareIt/server/dao/SharerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import fr.lyon.insa.ot.sims.shareIt.server.domain.Sharer;
 
 @Repository
-public interface SharerRepository extends CrudRepository<Sharer, Long>{
+public interface SharerRepository extends CrudRepository<Sharer, Integer>{
 
 }
